pyproject.toml
[project]
name = "harem-preprocessing"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "datasets>=4.1.1",
    "huggingface-hub>=0.35.0",
    "lxml==4.9.2",
    "python-dotenv>=1.1.1",
]
